@@ -11,18 +11,16 @@
     "prefer-stable": true,
     "minimum-stability": "stable",
     "require": {
-        "php": "^7.3",
-        "grizz-it/configuration": "^1.2",
-        "grizz-it/log": "^1.0",
-        "grizz-it/storage": "^1.0",
-        "grizz-it/vfs": "^1.0",
-        "php-unified/log": "^2.0",
-        "ulrack/aop-extension": "^1.0",
-        "ulrack/environment-extension": "^1.0",
-        "ulrack/invocation-extension": "^1.1",
-        "ulrack/kernel": "^1.3",
-        "ulrack/services": "^3.3",
-        "ulrack/tag-extension": "^1.0"
+        "php": "^8.0",
+        "grizz-it/configuration": "^1.3",
+        "grizz-it/log": "^1.1",
+        "grizz-it/services": "^1.0",
+        "grizz-it/storage": "^1.1",
+        "grizz-it/vfs": "^1.1",
+        "php-unified/log": "^2.1",
+        "ulrack/aop-extension": "^1.1",
+        "ulrack/environment-extension": "^1.1",
+        "ulrack/kernel": "^2.0"
     },
     "authors": [
         {
@@ -60,7 +58,7 @@
         ]
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.3",
-        "squizlabs/php_codesniffer": "^3.5"
+        "phpunit/phpunit": "^9.5",
+        "squizlabs/php_codesniffer": "^3.6"
     }
 }
